backend/requirements.txt
flask
snowflake-connector-python
flask-cors
python-dotenv
